Fix silly issues with tensorboardx. (#1210)
Tensorboard tries to interpret various things as arrays or histograms
that wandb would accept as unique scalars. This attempts to handle some
of these conditions and avoids crashing on the unexpected conditions.

e.g. Levanter likes to include things like:

layers_with_feature = ["x", "y", "z"]

surprisingly we can neither treat this as a scalar or an array: scalars
must be numeric, and arrays must be summable.